requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.10
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
environ==1.0
idna==3.3
PyJWT==1.7.1
pytz==2021.3
requests==2.26.0
sqlparse==0.4.0
typing_extensions==4.0.1
urllib3==1.26.7
